Move smart constructors to .Math
