+++ code/drasil-lang/Language/Drasil/ModelExpr/Math.hs
@@ -4,9 +4,11 @@ module Language.Drasil.ModelExpr.Math where
 import Prelude hiding (sqrt)
 import Control.Lens ((^.))
 import Language.Drasil.Expr (Completeness(..), DerivType(..))
+import Language.Drasil.ExprClasses (Express(express))
 import Language.Drasil.ModelExpr (ModelExpr(..), ($^), ($/), AssocArithOper(..),
-  LABinOp(..), VVVBinOp(..), UFunc(..), UFuncB(..), UFuncVN(..), UFuncVV(..), addRe)
-import Language.Drasil.Space (RTopology(..), DomainDesc(..), RealInterval)
+  LABinOp(..), VVVBinOp(..), UFunc(..), UFuncB(..), UFuncVN(..), UFuncVV(..), addRe,
+  AssocBoolOper(Equivalence, And), SpaceBinOp(IsIn), StatBinOp(Defines))
+import Language.Drasil.Space (Space, RTopology(..), DomainDesc(..), RealInterval)
 import Language.Drasil.Symbol (Symbol)
 import Language.Drasil.Classes.Core (HasUID(uid), HasSymbol)
 import Language.Drasil.Classes (IsArgumentName)
@@ -102,6 +104,10 @@ dbl = Dbl
 -- | Smart constructor for exact doubles.
 exactDbl :: Integer -> ModelExpr
 exactDbl = ExactDbl
+
+-- | Bring a Space into the ModelExpr.
+space :: Space -> ModelExpr
+space = Spc
 
 -- | Smart constructor for fractions.
 frac :: Integer -> Integer -> ModelExpr
@@ -218,3 +224,37 @@ sy x = C (x ^. uid)
 deriv, pderiv :: (HasUID c, HasSymbol c) => ModelExpr -> c -> ModelExpr
 deriv e c = Deriv Total e (c^.uid)
 pderiv e c = Deriv Part e (c^.uid)
+
+-- | One expression is "defined" by another.
+defines :: (Express a, Express b) => a -> b -> ModelExpr
+defines a b = StatBinaryOp Defines (express a) (express b)
+
+isIn :: Express a => a -> ModelExpr -> ModelExpr
+isIn a s@(Spc _) = SpaceBinaryOp IsIn (express a) s
+isIn _ _         = error "isIn target must be a Space"
+
+
+-- | Helper for creating new smart constructors for Associative Binary
+--   operations that require at least 1 expression.
+assocCreate :: Express d => AssocBoolOper -> [d] -> ModelExpr
+assocCreate abo [] = error $ "Need at least 1 expression to create " ++ show abo
+assocCreate _ [x]  = express x
+assocCreate b des  = AssocB b $ assocSanitize b $ map express des
+
+-- | Helper for associative operations, removes embedded variants of the same kind
+assocSanitize :: AssocBoolOper -> [ModelExpr] -> [ModelExpr]
+assocSanitize _ [] = []
+assocSanitize b (it@(AssocB c des):r)
+  | b == c    = assocSanitize b des ++ assocSanitize b r
+  | otherwise = it : assocSanitize b r
+assocSanitize b (de:des) = de : assocSanitize b des
+
+-- | Binary associative "And".
+andDEs :: Express d => [d] -> ModelExpr
+andDEs = assocCreate And
+
+-- | Binary associative "Equivalence".
+equivDEs :: Express a => [a] -> ModelExpr
+equivDEs des
+  | length des >= 2 = assocCreate Equivalence des
+  | otherwise       = error $ "Need at least 2 expressions to create " ++ show Equivalence
